Correction pour la class item
(ajout de speed en float)

+++ ForestSurvivor/ForestSurvivor/Shoot.cs
@@ -14,8 +14,8 @@ namespace ForestSurvivor
         private Texture2D _texture;
         private int _width;
         private int _height;
-        private int _x;
-        private int _y;
+        private float _x;
+        private float _y;
         private int _speed;
         private int _direction;
         private int _damage;
@@ -23,15 +23,15 @@ namespace ForestSurvivor
 
         public int Width { get => _width; set => _width = value; }
         public int Height { get => _height; set => _height = value; }
-        public int X { get => _x; set => _x = value; }
-        public int Y { get => _y; set => _y = value; }
+        public float X { get => _x; set => _x = value; }
+        public float Y { get => _y; set => _y = value; }
         public int Speed { get => _speed; set => _speed = value; }
         public Texture2D Texture { get => _texture; set => _texture = value; }
         public int Direction { get => _direction; set => _direction = value; }
         public bool Destroy { get => _destroy; set => _destroy = value; }
         public int Damage { get => _damage; set => _damage = value; }
 
-        public Shoot(int x, int y, int direction)
+        public Shoot(float x, float y, int direction)
         {
             _x = x;
             _y = y;
@@ -142,12 +142,12 @@ namespace ForestSurvivor
 
         public Rectangle GetShootRectangle()
         {
-            return new Rectangle(X, Y, Width, Height);
+            return new Rectangle((int)X,(int)Y, Width, Height);
         }
 
         public void Draw()
         {
-            Globals.SpriteBatch.Draw(Texture, new Rectangle(X, Y, Width, Height), Color.White);
+            Globals.SpriteBatch.Draw(Texture, new Rectangle((int)X,(int)Y, Width, Height), Color.White);
         }
     }
 }
